move local read setting dialog class
move it to BookProfileActivity
